fix(glial): widen synapse_id to uint64 (G7) — fixes 2M-scale collision
make_synapse_id used to be `pre * 10000 + post` returning uint32. Two
pre-existing bugs:
 (a) Overflows uint32 when pre_neuron_id >= ~429,497 (since pre*10000
     exceeds 4.29e9 there).
 (b) Collides deterministically when post_neuron_id >= 10000 — e.g.
     (pre=1, post=20000) and (pre=3, post=0) both map to id=30000, so
     assigning different astrocytes / microglia to them overwrites
     each other silently.

Flagged as critical C2 in the 2026-04-23 G1-G6 walkthrough. Tolerable at
current test scale but catastrophic at the 2M SNN target: astrocyte
modulation and microglia pruning would route to wrong synapses
deterministically.

G7 fix: bijective packing via `((uint64_t)pre << 32) | post`. Zero
collisions over any (uint32, uint32) pair. Cascading type widening:

- glial_integration_t synapse-keyed hash tables now use HASH_KEY_UINT64
  (neuron-keyed table stays UINT32 — neuron_id fits).
- id_list_t stores uint64 so both synapse_ids and neuron_ids share the
  same list type (neuron_ids widen trivially).
- Public API: glial_integration_assign_astrocyte_to_synapse and
  glial_integration_assign_microglia_to_synapse take uint64 synapse_id.
- microglia module widened through: monitored_synapse_t.synapse_id,
  monitored_synapse_ids array (UINT32_MAX → UINT64_MAX sentinel), and
  all microglia_*_synapse functions.
- Bio-async prune-request payload now reads 8 bytes (was truncating).

G7 regression test `SynapseIdBijectionAcrossOldCollisionPoints`
explicitly encodes the two (pre, post) pairs that used to collide in
the pre-G7 hash and confirms distinct astrocyte assignments survive.

All 9 unit + 5 integration glial tests pass.

+++ include/glial/integration/nimcp_glial_integration.h
@@ -229,7 +229,7 @@ nimcp_result_t glial_integration_set_spatial_neuromod_system(
  *
  * @param gi Glial integration system
  * @param astrocyte_id ID of astrocyte
- * @param synapse_id ID of synapse (pre_neuron_id * 10000 + post_neuron_id)
+ * @param synapse_id ID of synapse (G7: uint64, packed (pre<<32)|post)
  *
  * @return NIMCP_SUCCESS on success
  *
@@ -237,7 +237,7 @@ nimcp_result_t glial_integration_set_spatial_neuromod_system(
  */
 nimcp_result_t glial_integration_assign_astrocyte_to_synapse(glial_integration_t* gi,
                                                              uint32_t astrocyte_id,
-                                                             uint32_t synapse_id);
+                                                             uint64_t synapse_id);
 
 /**
  * @brief Assign oligodendrocyte to myelinate a neuron
@@ -267,7 +267,7 @@ nimcp_result_t glial_integration_assign_oligodendrocyte_to_neuron(glial_integrat
  */
 nimcp_result_t glial_integration_assign_microglia_to_synapse(glial_integration_t* gi,
                                                              uint32_t microglia_id,
-                                                             uint32_t synapse_id);
+                                                             uint64_t synapse_id);
 
 // ============================================================================
 // AUTOMATIC SPATIAL ASSIGNMENT
